requirements.txt
# Core web framework
fastapi
uvicorn[standard]

# AI and Google Cloud
google-generativeai
google-cloud-speech
google-cloud-texttospeech
genai-processors

# Database
motor

# Templating
Jinja2

# Utilities
python-dotenv
python-multipart
Pillow
